OPTLGMussoTest4/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary

# OpenTelemetry core
opentelemetry-sdk
opentelemetry-distro
opentelemetry-api

# Exporters
opentelemetry-exporter-otlp

# Instrumentations
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy

#Prometheus Client
prometheus_client

#Para la consulta a base de datos
asyncpg
databases